+++ src/Cmdty.Storage/LsmcValuation/LsmcStorageValuation.cs
@@ -117,6 +117,7 @@ namespace Cmdty.Storage
             ReadOnlySpan<double> endPeriodSimSpotPrices = regressionSpotSims.SpotPricesForPeriod(lsmcParams.Storage.EndPeriod).Span;
 
             int numSims = regressionSpotSims.NumSims;
+            double numSimsSqrt = Math.Sqrt(numSims);
 
             for (int i = 0; i < endInventorySpaceGrid.Length; i++)
             {
@@ -368,11 +369,11 @@ namespace Cmdty.Storage
             var pvBySim = new double[numSims];
 
             var deltas = new double[periodsForResultsTimeSeries.Length];
+            var deltasStandardErrors = new double[periodsForResultsTimeSeries.Length];
 
             Span<double> inventoryBuffer1 = returnSimInventory ? Span<double>.Empty : new double[numSims];
             Span<double> inventoryBuffer2 = returnSimInventory ? Span<double>.Empty : new double[numSims];
-
-
+            
             Span<double> thisPeriodInventories = returnSimInventory ? inventoryBySim[0] : inventoryBuffer1;
             Span<double> nextPeriodInventories = returnSimInventory ? inventoryBySim[1] : inventoryBuffer2;
 
@@ -383,6 +384,7 @@ namespace Cmdty.Storage
             int numTriggerPriceVolumes = 10; // TODO move to parameters
             var triggerVolumeProfilesArray = new TriggerPriceVolumeProfiles[periodsForResultsTimeSeries.Length - 1];
             var triggerPricesArray = new TriggerPrices[periodsForResultsTimeSeries.Length - 1];
+            var spotPriceTimesVolumeByPath = new double[numSims]; // Not memory efficient way to calculate delta standard error, but simpler code
 
             double forwardStepProgressPcnt = (1.0 - BackwardPcntTime) / periodsForResultsTimeSeries.Length;
             _logger?.LogInformation("Starting calculations of optimal decisions by simulation forward in time.");
@@ -418,7 +420,6 @@ namespace Cmdty.Storage
                 Day cmdtySettlementDate = lsmcParams.SettleDateRule(period);
                 double discountFactorFromCmdtySettlement = DiscountToCurrentDay(cmdtySettlementDate);
                 double discountForDeltas = lsmcParams.DiscountDeltas ? discountFactorFromCmdtySettlement : 1.0;
-                double sumSpotPriceTimesVolume = 0.0;
 
                 ReadOnlySpan<double> simulatedPrices;
                 if (period.Equals(lsmcParams.CurrentPeriod))
@@ -442,6 +443,8 @@ namespace Cmdty.Storage
 
                 double sumOverSimsInjectWithdrawVolumes, sumOverSimsCmdtyConsumed, sumOverSimsInventoryLoss, sumOverSimsPv;
                 sumOverSimsInjectWithdrawVolumes = sumOverSimsCmdtyConsumed = sumOverSimsInventoryLoss = sumOverSimsPv = 0.0;
+                double forwardPrice = lsmcParams.ForwardCurve[period];
+
                 for (int simIndex = 0; simIndex < numSims; simIndex++)
                 {
                     double simulatedSpotPrice = simulatedPrices[simIndex];
@@ -487,7 +490,7 @@ namespace Cmdty.Storage
 
                     double optimalCmdtyUsedForInjectWithdrawVolume = cmdtyUsedForInjectWithdrawVolumes[indexOfOptimalDecision];
 
-                    sumSpotPriceTimesVolume += -(optimalDecisionVolume + optimalCmdtyUsedForInjectWithdrawVolume) * simulatedSpotPrice;
+                    spotPriceTimesVolumeByPath[simIndex] = -(optimalDecisionVolume + optimalCmdtyUsedForInjectWithdrawVolume) * simulatedSpotPrice;
 
                     if (returnSimInjectWithdrawVolume)
                         thisPeriodInjectWithdrawVolumes[simIndex] = optimalDecisionVolume;
@@ -510,12 +513,12 @@ namespace Cmdty.Storage
                 double expectedInventory = Average(thisPeriodInventories);
                 storageProfiles[periodIndex] = new StorageProfile(expectedInventory, sumOverSimsInjectWithdrawVolumes/numSims,
                     sumOverSimsCmdtyConsumed/numSims, sumOverSimsInventoryLoss/numSims, sumOverSimsPv/numSims);
-                double forwardPrice = lsmcParams.ForwardCurve[period];
                 // Pathwise differentiation calculation makes assumption that simulated spot price is calculated as forward prices times some stochastic term.
-                // This is fine for the multifactor model in Cmdty.Core, but will not be the case for all models, e.g. a shifted lognormal model to account for 
-                // negative prices. TODO figure out best way to handle this, and/or document, or just abandon pathwise differentiation as delta calculation method
-                double periodDelta = (sumSpotPriceTimesVolume / forwardPrice / numSims) * discountForDeltas;
-                deltas[periodIndex] = periodDelta;
+                // This is fine for the multi-factor model in Cmdty.Core, but will not be the case for all models.
+                // TODO figure out best way to handle this, and/or document, or just abandon pathwise differentiation as delta calculation method
+                deltas[periodIndex] = spotPriceTimesVolumeByPath.Average() / forwardPrice * discountForDeltas;
+
+                deltasStandardErrors[periodIndex] = StorageHelper.StandardError(spotPriceTimesVolumeByPath, lsmcParams.SimulationUsesAntithetic) / forwardPrice * discountForDeltas;
                 progress += forwardStepProgressPcnt;
                 lsmcParams.OnProgressUpdate?.Invoke(progress);
                 lsmcParams.CancellationToken.ThrowIfCancellationRequested();
@@ -615,7 +618,7 @@ namespace Cmdty.Storage
             _logger?.LogInformation("Starting calculations of optimal decisions by simulation forward in time.");
 
             double forwardNpv = pvBySim.Average();
-            double standardError = pvBySim.StandardDeviation() / Math.Sqrt(numSims);
+            double standardError = StorageHelper.StandardError(pvBySim, lsmcParams.SimulationUsesAntithetic);
             _logger?.LogInformation("Forward Pv: " + forwardNpv.ToString("N", CultureInfo.InvariantCulture));
 
             // Calculate NPVs for first active period using current inventory
@@ -629,6 +632,7 @@ namespace Cmdty.Storage
             storageProfiles[storageProfiles.Length - 1] = new StorageProfile(expectedFinalInventory, 0.0, 0.0, 0.0, endPeriodPv);
 
             var deltasSeries = new DoubleTimeSeries<T>(periodsForResultsTimeSeries[0], deltas);
+            var deltasStandardErrorSeries = new DoubleTimeSeries<T>(periodsForResultsTimeSeries[0], deltasStandardErrors);
             var storageProfileSeries = new TimeSeries<T, StorageProfile>(periodsForResultsTimeSeries[0], storageProfiles);
             var triggerPriceVolumeProfiles = new TimeSeries<T, TriggerPriceVolumeProfiles>(periodsForResultsTimeSeries.First(), triggerVolumeProfilesArray);
             var triggerPrices = new TimeSeries<T, TriggerPrices>(periodsForResultsTimeSeries.First(), triggerPricesArray);
@@ -651,7 +655,8 @@ namespace Cmdty.Storage
                 _logger.LogInformation(Environment.NewLine + profilingReport);
             }
 
-            return new LsmcStorageValuationResults<T>(forwardNpv, standardError, deltasSeries, storageProfileSeries, regressionSpotPricePanel,
+            return new LsmcStorageValuationResults<T>(forwardNpv, standardError, deltasSeries, deltasStandardErrorSeries, 
+                storageProfileSeries, regressionSpotPricePanel,
                 valuationSpotPricePanel, inventoryBySim, injectWithdrawVolumeBySim, cmdtyConsumedBySim, inventoryLossBySim, netVolumeBySim, 
                 triggerPrices, triggerPriceVolumeProfiles, pvByPeriodAndSim, pvBySim, regressionMarkovFactors, valuationMarkovFactors);
         }